Write errors to S3
When processing a lambda management event and getting the remote config fails
write the lambda function name to S3 so that we can pick it up on the next
scheduled run.  This commit is only to write the errors to S3 and not handle
them, that will come in the next change.

I ended up writing a separate file in an errors directory for each lambda
function.  This way if someone is updating many lambdas at once and the remote
instrumenter fails for multiple of them, they don't accidentally overwrite
each other when writing a single file to S3.

Open Questions:
1. Right now this is for any error getting the remote config.  Should it be, or
should it specifically be the cache bypass limit exceeded error?

2. The contents of the error are being written to S3 now.  Do we care about the
specific error?

JIRA: https://datadoghq.atlassian.net/browse/SVLS-6102