Read errors back from s3
When instrumenting a function from a lambda management event fails, the
error will be written to an s3 object errors/{lambda function name}.json.
This commit lists all of the errors and parses the lambda function name out.
If there are any functions that failed on a lambda management event, the
scheduled invocation will see the s3 object and try to instrument the
function.

JIRA: https://datadoghq.atlassian.net/browse/SVLS-6102