Move instrument metrics into instrument function
Move emitting metrics for the front end from the same place instrument
was being called, so you would have to remember to call it, to inside
the instrument function.  This means that where ever the instrument
function is called from, the metric will be emitted and we don't need
to remember to do it.

I'm doing this ahead of adding a third place that would need this same
logic, so I pulled it out into the function before doing that change.

JIRA: https://datadoghq.atlassian.net/browse/SVLS-6102