Add linter and format files
Run the format command to format all of the code in the repo to the linter
standards, as well as adding a lint command to check if the repo follows
that standard.  I had to add in the jest globals so that describe, test,
etc don't cause issues.

My plan is to add an action (included in this PR) so that this runs when
new PRs are created.  From what I can tell on the github docs, this action
won't run on this PR but will in subsequent ones.  This should keep us honest
about code styling and felt like a good way for me to learn about using github
actions since I haven't before.  If there are better or different ways of
creating / testing actions please let me know!