Add unit test workflow
Add a workflow to check that the unit tests still pass on pull
requests.  I tried to make it so that it would only have to install
the dependencies once between this and the linter and reuse that,
but I couldn't find a way to do it.  It seems common in some of our
other packages to just re pull / install the package, but if someone
knows a way to avoid this I'm happy to adjust this.