Add integration test framework
Add the framework for us to do integration tests, along with an initial test.
I separated out the unit and integration test runs by using jest configs for
each of them so that they can be run separately.  The existing integration test
will assume a role and use that for accessing AWS resources.  The test that
is there now writes a file to S3 and invokes the remote instrumenter lambda
in the same way the 5 minute scheduled invocations do, then asserts some
things about the output.  To do this I had to also return the instrumentation
outcome from the lambda.

My intention behind having these tests is so that we can have tests close to what
we want customers to be doing, and because we will have those tests we will know
if the remote instrumenter is working or not.  These should be at the level of
"When a customer creates a lambda matching the remote instrumentation tags, that
function becomes instrumented" and not tied to any way that we implement this.

This is a bit of the first step towards doing more with this kind of testing.
The thing I would want to work towards (if we decide that this is the direction
we want to go) is towards increasing how automated this is.  Those steps roughly
include

1. Automating setup of these resources (as well as test specific resources) so that
you can run a single command and have a fresh test setup, and tear it down as easily.
2. Making the 4 config inputs automatically pulled in based on that setup
3. Run this all in a github action so that we know each PR will pass