Add setup and teardown scripts
Add scripts to publish a new layer version, setup the integration test
infrastructure, and delete the infrastructure after the tests.  I had
to make publishing the layer a separate script for now because of
how aws-vault is setup at different levels for that script vs the
rest.  I'll have to change that to work with different credentails
once I get this into a github action, but this gets everything working.

The test infrastructure consists of a role that the tests will use to run
as well as the remote instrumenter stack imported into CDK.  I used CDK
for this because we'll probably end up setting up a few similarly configured
lambdas and doing that will be much easier with CDK compared to having
to copy CFN.

I created a secret, Remote_Instrumenter_Test_API_Key, to stort the DD
API key for the remote instrumenter.