Run integration test in github action
Run the integration tests in a github action.  I had to pass the version
between the two jobs using an output since the file gets cleared between
jobs, but is required to build the stack.  It doesn't really matter if
the version matches there since it is destroying the stack, but it wasn't
hard to make the output.  I added an IAM user with permissions to do all
of the setup and teardown, as well as to assume the integration test role.

I still kept the actual tests using a role that is defined in the test
stack so that we can manage permissions there.

Also fix an issue with the template introduced in the last change.