Check list s3 contents before adding contents
The Contents section of the response isn't what I thought it was before,
it isn't part of the response if there is nothing there but I had thought
it was an empty list.  This was causing errors when there wasn't an error
stored in the s3.

Also fix the script to do setup / teardown since the import was causing
it to run setup again before the teardown.