Allow tests to run in any region
Setup the tests so that they can run in any region with no external setup.  They
can be run in a new region and will setup everything that they need to.  I had
to change one of the tests a bit to do that to check that a lambda that is created
after the remote config is setup is instrumented after some time.  I'll come back
in a later commit to add in all of the tests that we think were important.

Also move the version saving of the publish layers script to after the acknowledgement
so that if you start to run the script locally with the you don't write a version
that doesn't exist if you cancel.  Default the publishing script to version 1 if
one is not available instead of erroring out so that we can more easily run in new
regions.  Move this script into the setup script to make it easier to run and
configure the credentials.

Add the DD_PRESERVE_STACK tag to the stack because even though it is temporary, I
found it got deleted often while I was running tests against the version of the
stack.