Add deploy self monitoring action
Add the deploy self monitoring action to update our self monitoring stack when
a commit is pushed to prod.  This will keep the self monitor stack up to date
so we can continually run tests against it asynchronously.  To do this I had to
pick a region just for self monitoring, eu-west-3.  I picked this kind of at
random, if there are regions we prefer / don't prefer then just let me know
and it's easy enough to change.

Also included in this is some changes to make writing tests a bit easier by
putting the clients in getter functions to cache them and so that the tests
themselves don't need to initialize a client.