Add integration tests
Add tests for a few more cases we want to cover.  These include when a lambda
is updated with tags that mean it should be instrumented, the remote config
changes to include / exclude lambdas, and when the config updates a version
of the layers we use.

I also adjusted the retry strategy on the lambda client, with the remote
instrumenter running we tend to get throttled since that is also doing a
lot of lambda calls.  Having more retries and the adaptive strategy means
that the tests might take longer, but they're more likely to pass instead
of throwing exceptions about getting throttled.

Added the -i flag for jest to run the tests.  This makes jest "runInBand",
which runs the tests serially.  This is important because the remote config
is per the whole account, so tests that modify it will affect other tests.