Handle target function not existing
Handle the target function not existing, both in scheduled events and
lambda management events.  This can happen in a lambda management event
if someone creates / deletes a lambda before we process the event, which
is probably rare in practice but happens frequently during integration
testing.  Also handle this on scheduled events, where I noticed that the
instrumenter wouldn't clear the errors for the config changed events.