Skip error events
When a lambda management event comes in check if it was caused by an error
when deciding to skip it or not.  Theses kind of events include things like
400 errors when trying to update a lambda, and mean that no change was applied.
Since no change was applied to the lambda, we won't have to update anything.
I noticed this because it was causing an error because there is no response
element section of the event, and we depend on that for update events to get
the function name.