Generate function names in tests
Generate the function names based on the same naming seed as the config
during the tests.  I ran into an issue where I think Tal and I ran the
tests at the same time and I got resource conflict exceptions, so this
should prevent that.

This does not require you to regenerate your commit, it'll just be undefined
until you do and then the next run will have your name on the resources.