Handle CFN delete events
Handle a CFN delete event by uninstrumenting all of the lambda functions
that are in the account.  This happens very similarly to the scheduled
invocation workflow, except instead of getting the configs directly pass
in an empty array like there are no configs.  If anything fails to
uninstrument, send a response to CFN that the request failed.

For integration testing the cfn response module will callback whatever
url that gets passed to it, so create a presigned s3 url that goes to
our bucket that we can check if there is an object there or not to
simulate the CFN callback.

I did find I had to wait ~20 seconds for the everything to settle before
I could invoke the instrumenter with the CFN delete action, otherwise there
would be concurrent edit issues.  Even waiting for the lambda to finish
updating directly didn't work, I think there must just be a chain of events
that happen or something, not really sure.  Ideally the 20 second wait
wouldn't be there, so if you have ideas for an alternate check I'm happy to
try it out.

JIRA: https://datadoghq.atlassian.net/browse/SVLS-5122