Store and delete specific RCs by ID
Store the ID of a remote config we see when we get / set them.  Then between
tests we can delete the ones that are known, since there aren't likely to be
other RCs made in that span.  We still need to clear it fully beforehand and
afterwards though.

This will
cause the tests to do fewer calls to RC.  I noticed that sometimes the self
monitoring tests failed because they were throttled, so this will hopefully
help with that.