Handle stack creation events
When the stack is created, instrument the functions in that region just like
for a scheduled invocation event.  If anything fails just stop and report
success back to CFN since we don't want to fail stack creation because of
a problem instrumenting a function.

JIRA: https://datadoghq.atlassian.net/browse/SVLS-6259