Wait before trying to instrument
Sometimes a lambda will be updated and trigger the remote instrumenter to
run before the lambda exits the pending state.

I noticed this happening occasionally in the remote instrumenter self
monitoring, and hopefully this change makes that less frequent.

I also created a function to make a lambda client and let us get it from
where ever we need it to avoid having to pass it around everywhere.  It's
mostly a preference thing, so if we want to keep it passed in here I can
switch that.