Skip events from the remote instrumenter
Skip update events that originate from the remote instrumenter.  Previously
the instrumenter would instrument a function and trigger an update event,
causing the instrumenter to run again.  Without this check, the instrumenter
would get the remote config and do all the work to check if the function needs
to be instrumented or not.