Generate lambda function names for tests
Generate the lambda function name for each test so that it is unique and
not reused between tests.  I was seeing some resource conflict exceptions
and suspect that at least some of them were because we are re using the
lambda function name, and that is causing it to get updated more than
would be expected for a single test.  This should also make it a bit easier
to use, as a nice side effect.

Also wait for RC to be consistent when we need to for deleting the remote
configs which should make the clearing remote config test more reliable.