Remove DdRemoteInstrumentLayerAwsAccount parameter
Remove the DdRemoteInstrumentLayerAwsAccount parameter from the template so it isn't
confusing for customers who are trying to set the parameters.  This should always be
the prod account, but we need it to be our sandbox account for integration testing.

I switched the account number to be a mapping so it would be easy to change, then
loaded the yaml file in the test stack, changed the number, and saved the new template.

JIRA: https://datadoghq.atlassian.net/browse/SVLS-5784