Use secret arn instead of API key
Use a secret instead of an API to not have the API key in plaintext in the account.
This also means that the instrumented functions will have the secret arn instead
of the key directly, so the integration tests need to be updated.

JIRA: https://datadoghq.atlassian.net/browse/SVLS-6428