Assert no secrets are logged
After the integration tests run assert that no secrets are in the logs
from the integration test run.  This will only catch it if it is logged
as JSON, if someone does just console.log(secret) this will still pass.
I think that's okay because that is unlikely to happern by accident, the
more likely case is like what we encountered in the incident where we
logged something without realizing the secrets were in the object.

JIRA: https://datadoghq.atlassian.net/browse/SVLS-6429