Add gitlab pipeline
Define the shape of the pipeline for the remote instrumenter release process.  This will replace the GitHub action for deploying and running self monitoring.  This will deploy self monitoring, run the integration tests on it, then (eventually) validate that certain metrics meet whatever criteria we specify.  A good chunk of it isn’t done, but I wanted to do this PR first so that we can agree on what should be in the pipeline.

JIRA: https://datadoghq.atlassian.net/browse/SVLS-6305