Add publish scripts
Add the publish scripts, and some other minor tweaks to the pipeline.  This sets
it up so that it will check the version already published into S3 and only publish
if the version changes.  This means that in order for us to do a release, all we
have to do is change the version number on the template.  The layers will be deployed
and slotted into the template.

I set it back to version 0.3.4 since it would deploy to prod if we left it.

The prod version of publishing I did not try all the way, we may need to debug
on the first time we want to release but this should be very close to what
we need.

Also change the CW insights query for checking secrets to only include things with
5 characters or more, as well as a linting error for the template.

This does not (yet) create a release back in github.

Working sandbox deploy pipeline: https://gitlab.ddbuild.io/DataDog/serverless-remote-instrumentation/-/pipelines/58629759

Skipped instance: https://gitlab.ddbuild.io/DataDog/serverless-remote-instrumentation/-/jobs/844983148

JIRA: https://datadoghq.atlassian.net/browse/SVLS-6305