Add metrics validation step
Add the metrics validation step to gitlab.  This can be used to check
that the remote instrumenter was invoked, that it didn't error, and
that it didn't take too long.  This can be augmented with any custom
metrics we want to emit, if we want to.

I ended up having to do this as a separate test suite in typescript
since the Datadog API client seems to only be available in typescript.

I also made the prod publish not use the common rule for the gitlab
pipeline so that when someone changes things to test changes to
their pipeline they are less likely to accidentally deploy to prod.