Create release script
Create the release script to generate the commit updating the template
and includes the tag and push to a release branch.  Once that is there
we can add more of a description to the PR and merge it in.  This will
trigger the release pipeline to run since it is a new commit with
the template version being different.

JIRA: https://datadoghq.atlassian.net/browse/SVLS-6305