Invoke functions as part of instrumented check
Invoke the functions as part of the check that they are instrumented
correctly, they should still be invokable to make sure we don't do
anything to break customer functions.

Also make the prefix match both for permissions reasons and so that we
can see the invocations / errors from our dashboard.  Note that there
won't be anything on the dashboard now because I set it to only look
at the self monitoring versions.

We will probably have to expand how the function code is created for
when we have more languages supported, but that is something we can
tackle when we get there.

Dashboard: https://dd.datad0g.com/dashboard/csd-pma-aqw?fromUser=false&refresh_mode=sliding&from_ts=1742487216491&to_ts=1742573616491&live=true