Pull site and secret names to variables
Pull the site and secret names to variables.  Originally I was going to
change them from staging (datad0g) to ddstaging but that doesn't work
yet because of the permissions for remote config.  So I kept the values
the same but pulled them out like this to be easier to change when we
want to do that in the future.