Use sandbox account number for publishing to sandbox
Replace the prod account number mapping with the sandbox
account number when we deploy to sandbox.  Changing this is
what the integration test setup command does.

This is slightly hacky in that it isn't explicitly replacing
it like we can when we parse it to YAML, but should always work
because of the number prefix.