Copy layers from old region when adding a region
With the existing script when adding a new region we would have published
the newest version of the layer as the old versions.  IE, if there were 10
versions of the layer in an old region, adding a new region would have
published 10 layers in the new region but they would all have been the
same as the 10th version in the old region.  This change copies them
version by version.  This is kind of hard to test, I've ran each command
locally but we don't have new regions to publish to since we get them
dynamically