Add timeout for test custom resources
Add a timeout for the custom resource deploy time.  We run into this sometimes
when the custom resource doesn't return a response.  We encounter this more
than customers will, which is why I'm fine with doing it just for when we deploy
test stacks, though if other people feel differently let me know.  I ran into
this most recently because a role was deleted that was used by the test stack,
and so the lambda didn't get invoked.

AWS Docs: https://docs.aws.amazon.com/AWSCloudFormation/latest/TemplateReference/aws-resource-cloudformation-customresource.html#cfn-cloudformation-customresource-servicetimeout