Switch order of tag / instrument events
Switch the order of the tag and instrument events.  The current behavior is to
tag a function with the remote instrumenter tag after it has been successfully
instrumented.  This means that if the tagging fails, the function ends up in
a manually instrumented state, and the remote instrumenter won't recover from it.
By inverting the order, we make it so that even if the tag or instrument step fails,
the remote instrumenter will fix the issue on the next run.

Also added a check for the opposite case, where the untag fails and a function is
left tagged, but should not be instrumented.  The test confirms that the function
is correctly untagged after the next invocation.

I also ran into issues with the tagging test for 21 lambdas taking too long, I've
increased the timeout for this PR but will go back and do the checks in parallel
afterwards.