Parallelize and fix assertions for integration tests
Run the check for is function instrumented when there are 21 functions
in parallel, instead of sequentially.  This makes the test go from ~120
seconds to ~60.  When I ran this a few times locally I did see one
instance of a throttling exception, so I bumped up the retry limit.
Also changed this on the wildcard test and saw the duration change from
~35 to ~30, so a smaller diff but still an improvement.

Also remove the assertion that the instrumented functions need to be in
the response for the scheduled event.  Since there are a lot of lambda
management events also happening at the same time, even with setting the
RC moved to after the lambda creation we don't always see it in the
response because it was already processed.  This is okay because we
assert that the function is correctly instrumented right before this,
and if it got instrumented by the scheduled invocation or the lambda
event isn't really important, that it gets instrumented is.  This should
improve test reliability a lot when running the entire test suite,
which is what happens for integration tests as approval workflows or
the self monitoring tests in gitlab.