Add handling for failed tag events
Included in the tagging API response is a field for any functions that did
not get tagged.  This will continue to try to tag those functions 3 times
before throwing an error.  This does lead to potentially extra tagging calls,
since we're not adding them to existing batches that could support them, but
the calls are pretty quick so I think that's okay.