Re order tests
Re order the test pipeline in gitlab to run the lint / unit tests first,
then do the integration test setup.  This will take slightly longer overall
but in the case where a hypothetical totally-not-me someone starts a PR with
linter errors, the resolution is much faster since you won't have to wait
for the integration tests to run on a commit you're going to write on top of.

Sample run where I added `exit 1` to the unit tests: https://gitlab.ddbuild.io/DataDog/serverless-remote-instrumentation/-/pipelines/71897895