Delete config hash before invoking scheduled
I noticed a few times that the scheduled invocation integration tests failed
inconsistently around if the response included a skipped entry or not.  I
believe that this is happening because when a scheduled event happens
successfully, the instrumenter writes the config file to s3 and checks it
in future scheduled events.

Its hard to know for sure if this will fix the issue, but I'll keep an eye
out over the next few days to see if it continues to happen.

Some examples of tests failing that I think are caused by this (~1-2 per day)
* https://gitlab.ddbuild.io/DataDog/serverless-remote-instrumentation/-/jobs/1053000779
* https://gitlab.ddbuild.io/DataDog/serverless-remote-instrumentation/-/jobs/1053594325
* https://gitlab.ddbuild.io/DataDog/serverless-remote-instrumentation/-/jobs/1055652005