Setup retries when running in CI
We sometimes see intermittent failures when running integration tests, both in
CI and when running them locally.  Some are because RC staging occasionally
gives us 500s, and some is probably because scheduled events run every 5
minutes and our test suite takes longer than 5 minutes, so a scheduled event
run is going to happen at some point in our tests every time.  Where this happens
is going to be random since its based on when the stack was deployed and how
far along the tests are.

To try to make the CI tests smoother, allow them to be retried.

An example of a RC 503: https://gitlab.ddbuild.io/DataDog/serverless-remote-instrumentation/-/jobs/1055479846