Make function name a constant
Make the function name a constant so that if someone tries to deploy 2 instances
of the stack it will fail before the lambda got created.  I chose the lambda
to make constant like this because it is one of the first things to get deployed,
CloudTrail depends on this through the custom resource so it can't get deployed,
so the s3 bucket should remain empty (and be deletable) when this happens.

Also had to make the integration test infra change the constant, since we do end
up deploying multiple instances in a region for ease of development and it makes
it clearer when looking at a lambda what triggered its creation.