Add option for a user to use existing CloudTrail
Add the option for a user to use an existing CloudTrail.  We talked about having them
provide the ARN, but in a stackset where it could differ region by region and because
we don't actually need the ARN I opted to make it a true/false to make a stackset
deployment easier.  If anyone has different thoughts on that please share them.

Also turned the cloudtrail setting to not create one for us since we already have multi
region trails, we don't need to make our own.

JIRA: https://datadoghq.atlassian.net/browse/SVLS-7101